requirements.txt
matplotlib==3.7.0
numpy==1.24.2
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.10.1
tensorflow==2.11.0
tensorflow_datasets==4.8.0
tensorflow==2.11.0
tqdm==4.64.1
